refactor : entity, repository folder별 세분화
