+++ src/main/java/qna/domain/entity/User.java
@@ -1,11 +1,9 @@
-package qna.domain;
+package qna.domain.entity;
 
-import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import qna.UnAuthorizedException;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
 import java.util.Objects;
 
 @Entity
