+++ src/main/java/qna/domain/repository/AnswerRepository.java
@@ -1,6 +1,7 @@
-package qna.domain;
+package qna.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.entity.Answer;
 
 import java.util.List;
 import java.util.Optional;
