+++ src/main/java/qna/domain/repository/QuestionRepository.java
@@ -1,6 +1,7 @@
-package qna.domain;
+package qna.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.entity.Question;
 
 import java.util.List;
 import java.util.Optional;
