+++ src/test/java/qna/domain/entity/AnswerTest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.entity;
 
 public class AnswerTest {
     public static final Answer A1 = new Answer(UserTest.JAVAJIGI, QuestionTest.Q1, "Answers Contents1");
