+++ src/test/java/qna/domain/entity/QuestionTest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.entity;
 
 public class QuestionTest {
     public static final Question Q1 = new Question("title1", "contents1").writeBy(UserTest.JAVAJIGI);
