+++ src/test/java/qna/domain/entity/UserTest.java
@@ -1,4 +1,6 @@
-package qna.domain;
+package qna.domain.entity;
+
+import qna.domain.entity.User;
 
 public class UserTest {
     public static final User JAVAJIGI = new User(1L, "javajigi", "password", "name", "javajigi@slipp.net");
