+++ src/test/java/qna/domain/repository/DeleteHistoryRepositoryTest.java
@@ -1,10 +1,13 @@
-package qna.domain;
+package qna.domain.repository;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.domain.ContentType;
+import qna.domain.entity.DeleteHistory;
+import qna.domain.repository.DeleteHistoryRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
