+++ src/test/java/qna/domain/repository/UserRepositoryTest.java
@@ -1,10 +1,12 @@
-package qna.domain;
+package qna.domain.repository;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.domain.entity.User;
+import qna.domain.repository.UserRepository;
 
 import java.util.Optional;
